backend/requirements.txt
fastapi
uvicorn
websockets
python-multipart
jinja2
chartjs
pymongo
motor
python-dotenv
pandas
openpyxl
pymongo[srv]